Test packages now mirror source packageas
