+++ dsl/src/test/java/tech/pegasys/peeps/network/subnet/SubnetTest.java
@@ -10,11 +10,9 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.peeps.network;
+package tech.pegasys.peeps.network.subnet;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import tech.pegasys.peeps.network.subnet.Subnet;
 
 import java.util.ArrayList;
 import java.util.List;
