+++ dsl/src/test/java/tech/pegasys/peeps/node/model/NodeIdentifierTest.java
@@ -10,12 +10,10 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.peeps.node;
+package tech.pegasys.peeps.node.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import tech.pegasys.peeps.node.model.NodeIdentifier;
 
 import org.junit.jupiter.api.Test;
