+++ dsl/src/test/java/tech/pegasys/peeps/privacy/model/PrivacyAddressTest.java
@@ -10,12 +10,10 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.peeps.privacy;
+package tech.pegasys.peeps.privacy.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import tech.pegasys.peeps.privacy.model.PrivacyAddreess;
 
 import org.junit.jupiter.api.Test;
